Fix/option to clear (#14)
* [Fix] An attempt was made.

* [Fix] actually pass param now

* [Fix] actually pass param now

* [Fix] correctly detect whether a project/issue has changed

+++ src/main/kotlin/com/trackforever/main.kt
@@ -1,5 +1,8 @@
 package com.trackforever
+
+import com.trackforever.Application.Companion.arguments
 import com.trackforever.repositories.ProjectRepository
+import org.slf4j.LoggerFactory
 import org.springframework.boot.CommandLineRunner
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
@@ -9,10 +12,19 @@ import org.springframework.context.annotation.Bean
 class Application {
     @Bean
     fun init(projectRepository: ProjectRepository) = CommandLineRunner {
-        projectRepository.deleteAll()
+        if (arguments.contains("clearAll")) {
+            logger.debug("CLEARING ALL DATA")
+            projectRepository.deleteAll()
+        }
+    }
+
+    companion object {
+        val logger = LoggerFactory.getLogger(Application::class.java)
+        lateinit var arguments: Array<String>
     }
 }
 
 fun main(args: Array<String>) {
+    arguments = args
     SpringApplication.run(Application::class.java, *args)
 }
